+++ src/main.jsx
@@ -2,7 +2,7 @@ import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import App from "./App.jsx";
 import { BrowserRouter } from "react-router-dom";
-import store from "../store/store.js";
+import store from "./store/store.js";
 import { Provider } from "react-redux";
 
 createRoot(document.getElementById("root")).render(
